+++ laravel/documentation/changes.md
@@ -49,6 +49,7 @@
 - `Request::secure` now takes `application.ssl` configuration option into consideration.
 - Simplified the `paths.php` file.
 - Only write file caches if number of minutes is greater than zero.
+- Added `$default` parameter to Bundle::option method.
 
 <a name="upgrade-3.2"></a>
 ## Upgrading From 3.1